backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
motor>=3.6.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.1
pyjwt>=2.9.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.12
email-validator>=2.2.0